메인 배경 변경, json 파싱 테스트중 Test 에서 하고있음, put으로 업데이트와 delete 테스트해야함, post로 로그 넣기 테스트해야함

+++ Assets/Scripts/Data/UserData.cs
@@ -5,7 +5,7 @@ public class UserData {
     //public string uid { get; set; }
     public string did { get; set; } // 디바이스 id
     public string email { get; set; }
-    public long money { get; set; } // 가진 돈 단위 만원
+    public int money { get; set; } // 가진 돈 단위 만원
     public string rank { get; set; } // 직급 인턴 사원 대리 과장 차장 부장 이사 상무 전무 사장 회장
     public int heart { get; set; } // 하트 개수
     public int charge { get; set; } // 남은 충전 시간 초
